* commit 'e1f4157ab4fa5a1bc4098fb2bbe2c6dcb966185f':
  NTR Changed configuration of cache group cause karma doesn't support them
  NTR Switched title of the administration page
  NTR Use hardwired dependencies instead of semver
  NTR Added JS compression and minimized the output in the watch mode
  NTR Enhanced chunk splitting
  NTR Enhanced the error reporting when building the project
  NTR Fixed copy-after-build plugin which causes asynchronous operations on the filesystem

+++ src/Administration/Resources/administration/build/plugins/copy-after-build.js
@@ -17,7 +17,7 @@ WebpackCopyAfterBuild.prototype.apply = function(compiler) {
     const files = this._filesMap;
     const opts = this._opts;
 
-    compiler.hooks.done.tapAsync('sw-copy-after-build', (stats) => {
+    compiler.hooks.done.tap('sw-copy-after-build', (stats) => {
         stats = stats.toJson();
         const chunks = stats.chunks;
 
@@ -61,7 +61,10 @@ WebpackCopyAfterBuild.prototype.apply = function(compiler) {
                 }
 
                 fs.createReadStream(from).pipe(fs.createWriteStream(to));
-                fs.unlinkSync(from);
+
+                if (fs.existsSync(from)) {
+                    fs.unlinkSync(from);
+                }
             });
         });
     });
